+++ src/widgets/dialogs/switcher/SwitcherItem.cpp
@@ -1,0 +1,10 @@
+#include "widgets/dialogs/switcher/SwitcherItem.hpp"
+
+namespace chatterino {
+
+SwitcherItem::SwitcherItem(const QString &title)
+    : QListWidgetItem(title)
+{
+}
+
+}  // namespace chatterino
