+++ src/widgets/dialogs/switcher/SwitcherItem.hpp
@@ -1,0 +1,13 @@
+#pragma once
+
+namespace chatterino {
+
+class SwitcherItem : public QListWidgetItem
+{
+public:
+    SwitcherItem(const QString &title);
+
+    virtual void action() = 0;
+};
+
+}  // namespace chatterino
